+++ ModelChatIntent/schemas.py
@@ -1,0 +1,51 @@
+from datetime import datetime
+from typing import List
+from pydantic import BaseModel, Field
+from typing import Dict
+
+
+class TokenResponse(BaseModel):
+    token: str
+
+
+class Message(BaseModel):
+    text: str
+    clientId: str
+    sender: str
+    timestamp: datetime
+
+
+class MessagesList(BaseModel):
+    items: List[Message]
+
+
+class ChatResponse(BaseModel):
+    message: Message
+    history: List[Message]
+
+
+# Pydantic models
+class BlockTraffic(BaseModel):
+    goal: str = Field(description="it must be set to blockTraffic")
+    switch_id: int = Field(description="the id that identifies the switch")
+    ip_source: str = Field(description="the ip source address of the traffic to block ")
+    ip_dest: str = Field(description="the ip destination address of the traffic to block ")
+
+
+class LoadProfiling(BaseModel):
+    goal: str = Field(description="it must be set to setWeights")
+    switch_id: int = Field(description="the id that identifies the switch")
+    weights: Dict[int, int] = Field(description="the weights associated to the releted port as key ")
+
+
+class DeleteFlow(BaseModel):
+    goal: str = Field(description="it must be set to deleteFlow")
+    switch_id: int = Field(description="the id that identifies the switch")
+    ip_source: str = Field(description="the ip source address of the rule to delete")
+    ip_dest: str = Field(description="the ip destination address of the rule to delete ")
+
+
+class RateLimiter(BaseModel):
+    goal: str = Field(description="it must be set to setRate")
+    switch_id: int = Field(description="the id that identifies the switch")
+    rate: int = Field(description="the rate to apply on the switch")
